requirements.txt
requests
fastapi
uvicorn[standard]
fpdf2
